third_party/blink/web_tests/platform/mac-mac10.12/external/wpt/service-workers/service-worker/navigation-timing-extended.https-expected.txt

This is a testharness.js-based test.
FAIL Service worker controlled navigation timing assert_greater_than: workerStart marking should not wait for worker activation to finish expected a number greater than 1634043231901 but got 1634043231899.8
Harness: the test ran to completion.
